Add GameOutput class, use object in loop()
We no longer block on a readline() waiting for data from Q3.
The checkAndRead() method of GameOutput checks if data is available, reads it,
parses it into lines if possible, and returns. Worst case we end up waiting
for the select() timeout before continuing in loop(). (default 100ms)

This is important because now we can fireTick() on an approximately regular
schedule and not get hungup if we don't get a line from q3. It may not seem
like much but it gives us more consistancy in the timing of loop().