Database is now reliant on Peewee. Bans are functioning. Updated some things in the parser too the new system (lots more to go)
The base of the database system is in place, and right now I just need to start transitioning over the old parser code to the new system (uid is now cid, etc). After the parser is back up and running, a bit more cleanup around the house needs to be done, and then we're ready to re-add plugin support with some new nifty features. On schedule for a beta release next week.

Signed-off-by: Andrei Z <b1naryth1ef@gmail.com>